@@ -1,25 +1,20 @@
 namespace SoraTech
 
-open System
-open System.Collections.Generic
-open System.IO
-open System.Linq
-open System.Threading.Tasks
-open Microsoft.AspNetCore
 open Microsoft.AspNetCore.Hosting
-open Microsoft.Extensions.Configuration
-open Microsoft.Extensions.Logging
+open Microsoft.Extensions.Hosting
+open System.IO
 
 module Program =
     let exitCode = 0
 
     let CreateWebHostBuilder args =
-        WebHost
-            .CreateDefaultBuilder(args)
-            .UseStartup<Startup>();
+        Host.CreateDefaultBuilder(args)
+            .ConfigureWebHostDefaults(fun webBuilder ->
+                webBuilder.UseStartup<Startup>() |> ignore
+            )
 
     [<EntryPoint>]
     let main args =
-        CreateWebHostBuilder(args).Build().Run()
+        CreateWebHostBuilder(args).UseContentRoot(Directory.GetCurrentDirectory()).Build().Run()
 
         exitCode
